# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Course & Training Management System API (ИПР №1)",
    "version": "1.0.0",
    "description": "Безопасный REST API для системы управления курсами, личным кабинетом студента, кабинетом преподавателя и динамическим RBAC (Role-Based Access Control). Спецификация сгенерирована из декларативных аннотаций Swaggo (// Method Godoc, @Summary, @Description, @Tags, @Param, @Success, @Failure, @Router) в контроллерах."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 Base Endpoint"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Введите JWT Access токен (Bearer <token>)"
      }
    },
    "schemas": {
...
      "CourseResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "c-uuid-1"
              },
              "title": {
                "type": "string",
                "example": "Разработка на Go и TypeScript"
              },
              "description": {
                "type": "string",
                "example": "Практический интенсив"
              },
              "price": {
                "type": "number",
                "example": 45000
              },
              "start_date": {
                "type": "string",
                "example": "2026-11-01T10:00:00.000Z"
              },
              "max_seats": {
                "type": "integer",
                "example": 25
              },
              "available_seats": {
                "type": "integer",
                "example": 24
              },
              "status": {
                "type": "string",
                "example": "published"
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": {
            "id": "c-uuid-1",
            "title": "Разработка на Go и TypeScript",
            "description": "Практический интенсив по созданию масштабируемых систем.",
            "price": 45000,
            "start_date": "2026-11-01T10:00:00.000Z",
            "max_seats": 25,
            "available_seats": 24,
            "status": "published",
            "instructor": {
              "id": "instr-1",
              "specialization": "Golang, Highload",
              "user": {
                "full_name": "Алексей Архитекторов",
                "email": "alexey@instructor.local"
              }
            }
          }
        }
      },
      "CourseListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseResponse"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "example": 10
              },
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 10
              },
              "total_pages": {
                "type": "integer",
                "example": 1
              }
            }
          }
        },
        "example": {
          "success": true,
          "data": [
            {
              "id": "c-1",
              "title": "Golang Microservices & Docker",
              "price": 45000,
              "start_date": "2026-11-01T10:00:00.000Z",
              "available_seats": 18,
              "max_seats": 25,
              "status": "published"
            }
          ],
          "pagination": {
            "total": 1,
            "page": 1,
            "limit": 10,
            "total_pages": 1
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Операция успешно выполнена."
          }
        },
        "example": {
          "success": true,
          "message": "Операция успешно выполнена."
        }
      },
...
    }
  },
  "paths": {
...
    "/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Каталог курсов",
        "description": "Каталог опубликованных курсов с фильтрацией, пагинацией и поиском",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Номер страницы",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Количество на страницу",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "example": 10
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "description": "Поиск по названию или описанию",
            "schema": {
              "type": "string",
              "example": "Go"
            },
            "example": "Go"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "description": "Минимальная цена",
            "schema": {
              "type": "number",
              "example": 10000
            },
            "example": 10000
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "description": "Максимальная цена",
            "schema": {
              "type": "number",
              "example": 80000
            },
            "example": 80000
          },
          {
            "name": "instructor_id",
            "in": "query",
            "required": false,
            "description": "ID преподавателя",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.CourseListResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseListResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "c-1",
                      "title": "Golang Microservices & Docker",
                      "price": 45000,
                      "start_date": "2026-11-01T10:00:00.000Z",
                      "available_seats": 18,
                      "max_seats": 25,
                      "status": "published"
                    }
                  ],
                  "pagination": {
                    "total": 1,
                    "page": 1,
                    "limit": 10,
                    "total_pages": 1
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Создать курс",
        "description": "Создание нового курса (требуется особое разрешение courses:create)",
        "parameters": [],
        "responses": {
          "201": {
            "description": "models.CourseResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "c-uuid-1",
                    "title": "Разработка на Go и TypeScript",
                    "description": "Практический интенсив по созданию масштабируемых систем.",
                    "price": 45000,
                    "start_date": "2026-11-01T10:00:00.000Z",
                    "max_seats": 25,
                    "available_seats": 24,
                    "status": "published",
                    "instructor": {
                      "id": "instr-1",
                      "specialization": "Golang, Highload",
                      "user": {
                        "full_name": "Алексей Архитекторов",
                        "email": "alexey@instructor.local"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "403": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "Данные курса",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRequest"
              },
              "example": {
                "title": "Архитектура микросервисов на Go и Kubernetes",
                "description": "Глубокий практический курс по разработке распределенных отказоустойчивых систем на Golang.",
                "price": 45000,
                "start_date": "2026-11-01T10:00:00.000Z",
                "max_seats": 30,
                "status": "published"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Детали курса по ID",
        "description": "Возвращает полную информацию о курсе",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Course ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.CourseResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "c-uuid-1",
                    "title": "Разработка на Go и TypeScript",
                    "description": "Практический интенсив по созданию масштабируемых систем.",
                    "price": 45000,
                    "start_date": "2026-11-01T10:00:00.000Z",
                    "max_seats": 25,
                    "available_seats": 24,
                    "status": "published",
                    "instructor": {
                      "id": "instr-1",
                      "specialization": "Golang, Highload",
                      "user": {
                        "full_name": "Алексей Архитекторов",
                        "email": "alexey@instructor.local"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "courses"
        ],
        "summary": "Редактировать курс",
        "description": "Редактирование курса (разрешение courses:edit, только автор или админ)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Course ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.CourseResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "c-uuid-1",
                    "title": "Разработка на Go и TypeScript",
                    "description": "Практический интенсив по созданию масштабируемых систем.",
                    "price": 45000,
                    "start_date": "2026-11-01T10:00:00.000Z",
                    "max_seats": 25,
                    "available_seats": 24,
                    "status": "published",
                    "instructor": {
                      "id": "instr-1",
                      "specialization": "Golang, Highload",
                      "user": {
                        "full_name": "Алексей Архитекторов",
                        "email": "alexey@instructor.local"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "404": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "Обновляемые поля",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseRequest"
              },
              "example": {
                "title": "Архитектура микросервисов на Go (Обновленный)",
                "description": "Добавлены модули по gRPC, Kafka и OpenTelemetry.",
                "price": 49000,
                "start_date": "2026-11-15T10:00:00.000Z",
                "max_seats": 35,
                "status": "published"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "courses"
        ],
        "summary": "Удалить курс",
        "description": "Удаление курса (разрешение courses:delete, запрещено при наличии активных записей)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Course ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.SuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "success": true,
                  "message": "Операция успешно выполнена."
                }
              }
            }
          },
          "400": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "403": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/courses/{id}/students": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Список студентов курса",
        "description": "Возвращает список студентов конкретного курса (разрешение courses:view_students)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Course ID",
            "schema": {
              "type": "string",
              "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
            },
            "example": "c1a2b3c4-d5e6-7f8a-9b0c-1d2e3f4a5b6c"
          }
        ],
        "responses": {
          "200": {
            "description": "models.CourseStudentsResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseStudentsResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "course_id": "c-1",
                    "course_title": "Разработка на Go",
                    "total_students": 1,
                    "students": [
                      {
                        "enrollment_id": "enr-1",
                        "student_id": "usr-student-1",
                        "full_name": "Иван Иванов",
                        "email": "student@example.com",
                        "status": "confirmed",
                        "enrolled_at": "2026-10-04T02:00:00.000Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "403": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          },
          "404": {
            "description": "models.ErrorResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Неверные параметры запроса."
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}